Add utility to fill a device with 0xff
Can be used to erase all sectors on a NAND storage device without
incurring the additional writes from filling them with 0s.

Note this is not the same as a low level format, the NAND will not be
inherently aware that the blocks are no longer needed, though some may
figure it out.

Signed-off-by: Ian Munsie <darkstarsword@gmail.com>