dev_one & sdtest: Use synchronous IO & determine erase block size
Attempt to determine the erase block size for the storage media in use
from sysfs. We use that value to size the blocks to write at a time in
order to reduce the number of erase cycles that the card must perform.
This requires at least Linux v2.6.36 and direct access to the SD card
(ie, not through a USB card reader).

If we cannot determine the erase block size assume 8MB. 4MB is pretty
typical for modern cards and 8MB should give us a little extra breathing
room (It won't kill the card if we get it wrong, but it may wear it
unnecessarily).

Also use synchronous IO to increase the likelihood of the above modification
to work (otherwise the kernel may break up writes), as well as
increasing the accuracy of the displayed throughput and reducing cache
effects (which can be detrimental on devices with little memory, such as
the Nokia N900).

Signed-off-by: Ian Munsie <darkstarsword@gmail.com>