schroot: Support sudo without -E
Sigh, another moving part... Things like this remind me why I dislike
shell scripting...

The busybox sudo on the n900 has no -E flag (and always preserves the
environment). If the option exists we may well need to pass it in, but
if it doesn't the command will error out.

Test for it and pass it in if it is supported.

Signed-off-by: Ian Munsie <darkstarsword@gmail.com>