sdtest, dev_one: Allow start position to be specified
Allow the starting position to be specified for both of these utilities,
so that an interrupted run can be resumed without having to wait for the
entire device to be read/written again.

Note that passing in the start position to sdtest will use it for both
the read and write phases - so it isn't completely suitable to resume a
run that was interrupted during the write phase (but that can be done
by running the read phase again over the whole device).

Signed-off-by: Ian Munsie <darkstarsword@gmail.com>