sdtest: Make random pattern block size agnostic
Previousely the random pattern we wrote to a device would vary depending
on the erase block size of the device. Since we use a default block size
of 8MB if we can't determine it then we could write one pattern to a
device plugged directly into a (non-USB) SD reader, then fail the verify
if we read it with a USB SD reader between us (or an older kernel,
or...) since the pattern we generated would be different.

We now seed the random number generator every 4K instead of every
BLOCK_SIZE so that variations in the erase block size will not change
the resulting pattern.

Also, while we are changing the pattern anyway reduce the number of
calls to rand() to reduce the CPU usage on embedded devices.

Signed-off-by: Ian Munsie <darkstarsword@gmail.com>