sdtest: Reduce number of rand() calls
Ok, this is a little more tricky, but again the goal is to reduce the
number of times we need to call rand() - which is a bottleneck on some
embedded CPUs.

rand() gives us 31 bits of random data, so almost 97% of 4 bytes. While
it would be nice and simple if it gave us 32 bits the fact is that it
doesn't (boo).

So, let's use the full 31 bits that rand() does give us, and use a
single bit from a second rand() call for the 32nd bit to make up four
whole bytes of random data.

So, instead of a call to rand() for every byte, we now have one call for
every four bytes, and an additional call every 124 bytes (4*31) to get
us those last few missing bits.

Signed-off-by: Ian Munsie <darkstarsword@gmail.com>