schroot: Allow some parameter checks only die() if we are root
It is possible to enter a chroot that we can't see (due to filesystem
permissions) until we are actually root, so some of the tests may raise
a false-negative and abort the process before we become_root() when they
would have passed.

Make these tests only display a warning if they are fail as a limited
user, but die() if they fail as root.

This also has the additional advantage of allowing commands run from the
schrootrc (such as a mount command) to be able to run as root before we
can abort the script with these checks.

Signed-off-by: Ian Munsie <darkstarsword@gmail.com>