Start tracking common shared config files
There are a number of config files I tend to use on every machine, and
they have all diverged from each other. This can happen a number of ways
- either by creating them from scratch each time (.vimrc especially),
copying them from an older version which happens to be more convienient
to access at the time, or modifying them for one machine and never
propergating the change back.

It's time I did something about it.

The desktop directory is intended to start collecting these scripts and
generalising them to work on all of my machines. Where they can't be
generalised any machine specific customisations can be tracked in an
alternative branch.

For the moment add the .Xresources and .vimrc files from one of my
laptops. The .vimrc may need some fixes to work with other machines -
I'll do that as I come across the issues. I also desperately need to
start tracking my wmii scripts soon.

Signed-off-by: Ian Munsie <darkstarsword@gmail.com>