fish vi-mode: Add 'vi_mode' variable, fix overwrite mode, bind_all hacks
The variable vi_mode will now be set to reflect the current mode.

Unfortunately it seems there is some magic that causes the prompt to
only repaint if the default binding is set to self-insert (it turns out
that this same magic allows overwrite mode to work)...

So, we now bind all the printable 7-bit ASCII symbols to do nothing in
normal mode, and make the code we were using for replace + overwrite
modes generic enough so we can share it.

Signed-off-by: Ian Munsie <darkstarsword@gmail.com>