fish vi-mode: First step towards getting directions working
Modify the W,E & B commands with directions in mind.

This works out the way to make the embedded python and fish shell
interract reliably (I hope!), such that the python can manipulate the
commandline and cursor arbitrarily.

I'm going to do most of the processing in Python - originally I was
thinking about returning three strings from Python (before the bit to be
manipulated, the bit from the cursor to the destination of the
direction, and the rest), but that would have been difficult as any of
those strings could be multi-line and I couldn't think of a good way to
reliably separate them.

Fish passes the command, direction, cursor position and commandline to
python in it's argv. Python returns the new commandline and position to
fish via stdout - the last line of which is the new position, everything
else is the new commandline.

PLEASE NOTE that this removes the hack I used to calculate how many
escaped characters were in the commandline, meaning this will break on
the latest official release of fishfish, and a recent development build
is required. This is really unavoidable - to get directions working I
need to manipulate the commandline directly rather than just asking fish
to delete-forward (etc), so there is little point trying to support
older builds.

Signed-off-by: Ian Munsie <darkstarsword@gmail.com>