fish vi-mode: Implement w, e, b, 0, _, h, l as directions
Lots of stuff works, de can now be used to delete to end of word, c0
will change to the start of commandline, etc.

The lower case versions of w, e and b are slightly too inclusive, but
writing those regular expressions was doing my brain in, so these will
do for now (they seem pretty close to me :)

0, $ and _ don't strictly do the right thing - they go to the start/end
of the whole command line instead of the start/end/first non-whitespace
of the current line.

Directions that place the cursor to the left of the target now specify
an offset as -1 which is only applied when moving the cursor, not when
using them as the target of a command.

Signed-off-by: Ian Munsie <darkstarsword@gmail.com>