fish vi-mode: Implement vi-style single level undo
Just single level support for now, multi-level vim style is a lower
priority for me (also, presumably I'd put them in an array, but I'm not
sure how that would work when a commandline could potentially be
multi-line, need to check if I can do arrays of arrays in fish).

Undo history is not cleared on execute for now. If we later support
multi-level undo history we should definitely do this, but with single
level it doesn't matter so much.

Signed-off-by: Ian Munsie <darkstarsword@gmail.com>