fish vi-mode: Enhancements by Maxim
Hi,

I've updated python's 'print', added colors for mode indicators and added calling vi_mode_user in the end of
functions, which set vi_mode. You can check the code in the attached file.

By the way, prefixing functions with the same prefix will simplify splitting the vi-mode.fish in separate files and adding it to the fish.