fish vi-mode: Fix broken directions
The same string that we were setting to the vi_mode was depended on in
the python direction_command function, so allowing that to be changed
by the user caused issues.

Instead still pass the same string around the script (but now expanded
out to a full word), but allow the string displayed in the vi_mode to be
defined by the user.

Signed-off-by: Ian Munsie <darkstarsword@gmail.com>