fish vi-mode: Clarify usage instructions
I've been sourcing the file from config.fish, but just assumed that it
would work so long as fish could find it. It seems that may not actually
be the case, so explicitly mention it needs to be sourced.

I guess it really depends on the load order of everything, which I have
not looked into.

Signed-off-by: Ian Munsie <darkstarsword@gmail.com>