fish vi-mode: Implement ge direction
Like b, this reverses the commandline and searches with the same regexp
as w.

This is our first g direction and works out how directions are handled.
In the generic bind_directions we bind g to bind_directions_g which in
turn binds the g directions.

The g *commands* are currently bound elsewhere by __vi_mode_g (which
also calls bind_directions_g for moving the cursor). I'm not sure if
this is a good idea to have these separately, it might be better to
rename bind_directions_g to bind_g and do everything from there (since
bind_directions is always called in normal mode).

Signed-off-by: Ian Munsie <darkstarsword@gmail.com>