fish vi-mode: Add ~ g~ gu and gU commands
Pretty simple implementation, but there is a caveat that thanks to
fish's automagic completion the results of these commands will not
always be immediately visible (ie, the commandline is changed correctly,
but still displays the case from the command that automagic completion
wants to use).

Ideally I'd disable automagic completion altogether whenever we are in
any mode except for insert as it doesn't really make sence.  I need to
see if there is a way to do that or send in a feature request/patch.

Signed-off-by: Ian Munsie <darkstarsword@gmail.com>