fish vi-mode: Suppress autosuggestion when entering normal mode
Fish now has a command to allow scripts to temporarily suppress the
autosuggestion feature. We use this when entering normal mode as the
autosuggestion is a bit out of place in that mode, and can also
interfere with the displayed result of the change case commands (they do
change the case on the commandline, but that might not be displayed if
an autosuggestion is active).

Signed-off-by: Ian Munsie <darkstarsword@gmail.com>