fish vi-mode: Special case cw and cW when cursor is on non-space
To match the special meaning of cw/cW in vim (i.e. "change word", rather
than change to start of next word).

As per the comment in the code, in practice this still doesn't quite
match vim's implementation of the special case when the cursor is on the
last character in a "word", for vim's definition of a word.

Reported-by: Michael Ellerman <michael@ellerman.id.au>
Signed-off-by: Ian Munsie <darkstarsword@gmail.com>