Add git-push-wd and inotify_sync scripts
Combined together these can be used to automatically synchronise a
remote git tree with a local one whenever it is changed (including dirty
modifications).

Add this to .git/config on the remote tree:
[receive]
        denyCurrentBranch = warn

Then run inotify_sync.sh <remote> in a terminal in the root of the local
git tree and any time a file is written it will automatically push the
state of the working tree to the remote (on a branch called
"last-deployed") and check it out.

Signed-off-by: Ian Munsie <darkstarsword@gmail.com>