blue-tether: Steps to better track DHCP client
After the connection has been up for a while, it's IP address
mysteriously disappears (this also happens with blueman). I strongly
suspect that this issue lies on the side of the android tablet not
responding properly to a IP renew request, but I need to examine what
actually happens when the address goes away to be sure.

Add the -d flag to dhclient to keep it in the foreground, which will
hopefully make it print out what happens when it does a renew and might
shed some light. If that doesn't work I'll have to fire up wireshark.

Also put the kill/wait in a finally block to be sure the DHCP client
won't remain running in exceptional circumstances if the script is
killed.

Signed-off-by: Ian Munsie <darkstarsword@gmail.com>