fish vi-mode: Minimal repeat last command functionality
Currently this will only repeat the functionality that the
__vi_mode_direction_command performs.

In practice, this means that commands that remain in normal mode (such
as de) will mostly work, but functions that change the mode (such as ce)
will not.

The main thing I use . for is to repeat delete word commands, so this is
enough to get started.

Signed-off-by: Ian Munsie <darkstarsword@gmail.com>