Import correctcase.py (mp3 English Title Capitaliser)
This is something I wrote a few years ago and was one of the first
pieces of non-trivial python code I wrote. It's been sitting on my old
laptop nagging at the back of my mind that it was still missing from my
junk code.

It goes through an mp3 music collection and corrects the capitalisation
on all the artist, album and title fields. It tries to handle a bunch of
edge cases and always asks for confirmation before making changes, so it
should be pretty safe to use.

I guess the main thing it's missing is ogg & flac tag support and
perhaps a better UI.

Signed-off-by: Ian Munsie <darkstarsword@gmail.com>