Rename desktop directory to config
I've been thinking about doing this for a while now... With more and
more things being added I think it's probably important I do this now
before too many more symlinks and URLs point to the old directory.

Create a desktop symlink for compatibility, to be removed as soon as I
update all the external stuff that points here.

Signed-off-by: Ian Munsie <darkstarsword@gmail.com>