blue-tether: Track connection state & support automatic reconnecting
This now keeps track of the bluetooth PAN connection, starting the DHCP
client when it comes up and killing it when it goes down.

If a reconnect interval is specified with -r or in the config file, it
will automatically attempt to reconnect at that interval if the network
goes down. If no interval was specified it will terminate.

Signed-off-by: Ian Munsie <darkstarsword@gmail.com>