Introduce git-am-tool
Every time I apply a patch in git that doesn't quite apply cleanly I
always end up doing a convoluted dance and more often than not mess
something up and have to start over.

This script is my attempt at streamlining this process - the idea is to
provide something like git mergetool, but for applying patches where we
don't know or have the version of the file that the patch originally
applied to.

For every file that git-apply fails to patch cleanly we open an editor
with the file on the left opened to where the first hunk was supposed to
go and the rejected hunks on the right. Once we have edited the file to
manually apply the failed hunks the file is added to the index and the
reject file is deleted.

Signed-off-by: Ian Munsie <darkstarsword@gmail.com>