fish vi-mode: Fix cw/cW corner case
The vim doco indicates that cw and cW are treated the same as ce and cE,
however there is one case where they behave differently - specifically,
when the cursor is on the last character in a word or on certain
punctuation.

This patch should replicate vim's behaviour in that corner case in most
circumstances (we still don't handle multiple punctuation symbols in a
row quite right).

Signed-off-by: Ian Munsie <darkstarsword@gmail.com>