XResources: Cleanup and add Shift+KP_Insert as extra paste button
My new laptop doesn't have a normal insert key, but it does have a
keypad with an insert key on the 0, so add that as an extra paste
button. I'm adding both KP_Insert and KP_0 so I don't have to worry
whether num lock is on or off.

While I'm at it, clean up the file by matching on the VT100 widget
rather than the XTerm/UXTerm name (*XTerm doesn't work).

Signed-off-by: Ian Munsie <darkstarsword@gmail.com>