fish vi-mode: NOTE RENAMED FUNCTION CAUSING BREAKAGE
Caps to try to get your attention to read this ;)

On 2012-10-21 fish renamed fish_user_keybindings to
fish_user_key_bindings and dropped backwards compatibility support for
the old name.

THIS MEANS YOU NEED TO EDIT YOUR ~/.config/fish/fish.config AND RENAME
THE FUNCTION THERE!

Edited the comment in the file to note the old name is deprecated.

Signed-off-by: Ian Munsie <darkstarsword@gmail.com>