git-am-tool: Allow specifying patch & optionally ignore untracked files
This way the script can be used when we only have a simple patch file
and not something that git-am can process.

The option to ignore untracked files is because I now have a per-tree
build.sh script in a lot of my git trees that would not be appropriate
to commit since it just automates some of my local build setup that
doesn't apply to anyone else.

Signed-off-by: Ian Munsie <darkstarsword@gmail.com>