Add an old python TTY UI class I wrote
The original idea here was that I would create a consistent API that
could be used to abstract whether the app was running in in a TTY or had
curses to do things like displaying coloured text, asking a y/n
question, etc.

I never got around to writing the curses implementation and never
stabilised the API, so this just ended up being a quick n' dirty way to
add some fancy output to some scripts.

I'm about to use it again, so I'll add it here.

If I ever get around to working on this again expect the API to change.

Signed-off-by: Ian Munsie <darkstarsword@gmail.com>