Track some Arduino sketches
These are a couple of sketches I've written for the LeoStick Arduino
over the past year.

press_me is basically a glorified hello world for an LCD display and a
button and was one of the first sketches I wrote for the LeoStick.

_7seg contains a library to draw arbitrary text (with a font of my own
creation) or other test patterns on a 4 digit 7 segment display,
optionally using a serial to parallel shift register to reduce the
number of pins on the arduino that have to be connected to the display.
One of it's test patterns also operates a stepper motor - hey, I didn't
want to pull everything off my breadboard ;-)

logicgate is a sketch I wrote while I was learning how to build logic
gates out of diodes and transistors. Attach the A and B PINs to the
inputs of a logic gate and the result pin to the output and it will
display what kind of logic gate it is on a seven segment display (using
the code from _7seg - I really need to re-split that back out and
properly librarify it).

remote is a work in progress to control some RF remote controlled
power boards / sockets using a 433.92MHz ASK transmitter. It can already
work with the dimmers from Aldi, I still need to hook up an oscilloscope
to my Heller remote for the power boards to capture their codes.  I also
plan on making it receive IR codes and relay them to the computer.

Signed-off-by: Ian Munsie <darkstarsword@gmail.com>