Introduce wget2axel.py
This script will create an axel state file for a file that has already
been partially downloaded with another application to allow the
remainder of the file to be downloaded accelerated. It should work with
partial files from almost any *non-accelerated* downloader, such as
wget, curl, firefox, etc.

The idea is that if you have started a download outside of axel, and
later on realise that the download is going too slow and will take too
long to complete that you can transfer the remaining part of the
download to axel, without having to re-download the part of the file
that you already have.

The state file that this script creates will tell axel that it was using
a lot of connections, calculated such that there will actually only be
four or five connections in use to download the incomplete part of the
file (the remaining "connections" are already complete as they cover the
sections of the file that have already been completely downloaded).

This script has some limitations (http only, can't follow redirects, no
ability to override the referrer or user agent, etc.) and until it has
undergone more testing I would suggest only using it on files where you
can verify the checksum afterwards.

Also, this script isn't magic - if the server doesn't support resuming a
partially downloaded file this won't help you, so you might want to
check that before you stop your in progress download in case you have to
start again from the beginning :)