i3 config: Add default generic config, add thinkpad specific bar
If gen_config.sh is involked with no arguments, it will now combine
the "base" and "generic" fragments. "generic" will hold anything that
can't be duplicated in the config file and therefore must be in each
individual configuration if it is customised anywhere.

Specifically, add an i3status config for my thinkpad that shows the free
space in /home - since many machines just use one filesystem for / and
/home, make the eee config the default generic one.