Add gg.fish and rg.fish (recursive grep)
rg.fish is a wrapper around recursive grep to make it behave more like
git grep. Specifically, if it is being output to a tty it enables colour
highlighting and pipes it into less with the -F flag to have less quit
if there is less than a screen of output and no wrapping.

gg.fish is an alias around git grep whenever it is run within a git
tree. If it is run outside of a git tree it falls back to rg.git