i3companion: Add libpurple -> libnotify notification redirector
Q: "But DSS, there are already libnotify plugins for pidgin, why do you
need this?"

A: pidgin-libnotify doesn't support finch and purple-libnotify+ crashes
finch.

For this to work, all you need to do is enable the "DBus Example" plugin
in Finch (Alt+A -> Plugins) or another libpurple client and start
purple.py (configs/base starts it automatically at the moment). It
subscribes to the ReceivedChatMsg and ReceivedImMsg signals from
libpurple and simply redirects them to libnotify.