i3 config: Add mod+q to find window by title
Because I can never remember which tag I had my useful.otl file open on.

I've disabled the focus child shortcut that was on this keybinding since
I've never used it (besides which it was disabled in the default
config). In wmii I had this implemented by an action with mod+a, and I
might end up doing something similar...