check_acf.py: Add option to --move extra files out of the way
This is useful to revert a game to it's pristine vanilla state without
completely blowing any extra files away. e.g. to restore a heavily modded game
back to vanilla, or for when you aren't sure if you need the extra files or
not.

The extraneous files will be moved to a directory tree under
SteamApps/common/<Game>~EXTRANEOUS

There is no undo function in the script yet, however rsync should work just
fine to merge them back in if needed.