i3 config/thinkpad: Add clean-xterm-log.sh
I'm experimenting with enabling xterm logging globally on this box, and
this script is run periodically from cron to remove log files that are a
couple of days old. I have this line in my crontab:

0 * * * * ~/.i3/local/thinkpad/clean-xterm-log.sh

To enable logging globally, I also need this line in the i3 config which
I have not committed yet (there does not appear to be an Xresource or
similar to do this):

bindsym $mod+Return exec i3-sensible-terminal -l

I don't really want to pull that out of the base config since this is a
very unusual change and adding it to the thinkpad config causes i3 to
complain about a duplicate binding.  I'll probably end up expanding
gen_config.sh to allow lines in the base config to be overridden somehow
(perhaps via a sed script)... TBD