Steam trading-cards.py: Catch some games that were missed
It seems in some circumstances the appID from the users all game list
doesn't match the IDs we got from the store search.

In particular, War of the Roses: Kingmaker shares trading cards with War
of the Roses but wasn't showing up.

This now queries Steam for the details of every single unmatched game
(25 at a time) to find these outliers.

At the moment I'm using the fullgame property to match up the IDs -
that properly is meant for demos, so I don't know if this is strictly
correct, hence I mark these as 'D' in the owned column.