Steam trading-cards.py: Search for all categories (e.g. DLC)
War of the Roses: Kingmaker wasn't showing up as owned since it is DLC
which we weren't searching for (the base game is F2P, so a situation can
arise where Kingmaker is Owned, but the base game is not).

This isn't really ideal as this means DLC and (worse yet!) Demos will
clutter up the list. I'm going to work on filtering out demos at least
since they are not valid for cards.