i3companion: Add screen backlight control via RANDR
This is necessary on my Macbook Air which doesn't handle this in
firmware. I have been running gnome-settings-daemon to fill this void,
but I keep getting annoyed at it for various reasons.

Using RANDR to set the backlight has the advantage that I don't need to
be privileged or insist on installing any kind of suid root helper.

The actual function is a bit of a mess and needs to be refactored -
partly because it uses xcb to talk to X asynchronously, and partly
because I was figuring it out as I went.