i3companion: Add keyboard backlight control via UPower
Well, that's the last thing I needed gnome-settings-daemon for -
***UNINSTALL***    Ahh, much better!

UPower provides a DBus interface to control the keyboard backlight,
avoiding the need to be privileged or requiring manual installation of a
suid root helper.

If this doesn't work and you try switching your display manager to gdm3.
To be honest I'm not actually sure what starts up UPower in the
background - I'm assuming gdm3, but that's really just a guess since I
haven't bothered checking. Meh, works for me.

I also notice UPower provides an interface to suspend & hibernate, so
I'll probably hook something into that in the near future. Though, what I
really need is a way to override the power button to ask what to do (and
notably, I'd really like a way to shut down without root, not just
suspend - I'll have to figure out how Gnome & KDE do it).