i3companion: Handle Power Button for halt/reboot/suspend/hibernate/logout
This is a bit gross - Debian has some magic scripts that handle the
power button. They try to detect if their is some other power management
daemon running that should take care of it for them, but unfortunately
provide no means to register such a daemon with them :(

Instead they look for a bunch of well known daemons that might be
running - and mine is not particularly well known - at least not yet :-P

So, part of this patch creates a fake "gnome-settings-daemon" - it's
just a simple C program that does nothing (other than print out an
explanation), which satisfies part of the test. For the rest I register
just enough of the org.gnome.SettingsDaemon.Power DBus interface in
i3companion.

If you want to use this you will need to build the fake
gnome-settings-daemon in the i3companion directory, otherwise pressing
the power button will probably generate a Python Exception (I'll fix
that later).

Pressing the power button will now bring up a menu asking whether to
shut down, reboot, suspend, hibernate or log out. Considering the
retarded placement of the power button on my Macbook Air this is a huge
step up from shutting down the computer if my finger missed the
backspace key (grrr). I even added an "OOPS" option which is selected by
default in case of two slip-ups in a row.

I should really check if I have permission to do those things, but hey -
this is *MY* system, if I have denied myself one of those things it's
only because it doesn't actually work anyway... Admittedly, that's still
an area that needs some serious fixes on my Macbook Air, but one step at
a time...