Add rdiff-backup over reverse-encfs scripts
Uses reverse encfs to create an encrypted view of the directory to be
backed up, then runs rdiff-backup to back it up to a remote server.

backup-wrapper.sh is also provided, which is intended to perform the
backup from a cron script. To use it the password must be stored in a
file (see the comment in the script). It will attempt to find and use a
running ssh-agent to be able to log into the remote backup server, so
will only work if the user has run ssh-add prior to the script running.