i3/randr: Fix backlight adjustment to work with more drivers
Use all four bytes from the backlight integer from randr to work with
drivers that have a backlight range of >= 255. It would be nice if
python-xcb did this automatically (or am I just doing something wrong?),
but at least now I understand why I got multiple values in the first
place.

Also change the default backlight deltas to 5% instead of 1 unit - the
intel_backlight driver has a range of 4437, so bumping that 1 unit at a
time is not very useful.