steam/check_acf.py: Add support for alternate steam libraries
We were previousely expecting the depotcache directory to always be in the
library directory along with SteamApps, but when Steam added support for
multiple libraries that became no longer the case. The depotcache directory is
always in Steam's root, which is not necessarily the same as the library root.

This tries to locate Steam's root in a number of ways - if the library root has
a directory called 'depotcache', that is assumed to also be the steam root. On
Windows/cygwin it reads the steam path from the registry, falling back to
searching some common paths (which may easily happen on cygwin since the
cygwinreg python module is unlikely to be installed).

A command line flag (--steam-root=...) can be used to override this auto
detection if it fails.