Add MIT license file
Doesn't mean I wouldn't appreciate patches, but given the nature of this
repository you can do whatever you like with it.

To be clear - this repository was always intended to be liberally
licensed (it started as random code snippets on my blog), so consider
this retrospective to all commits made by myself in this repository.