spotify: Rip out ad-muting code
Spotify no longer reports that it is playing ads through d-bus - it
now claims to be playing the previous track.

I have identified an alternate means to determine that it is playing
an ad, but recently it's ads haven't been bothering me as much as they
used to (they've stopped Biebering me at least) so I'll leave that for
another day.