Add steam/library_sync.py
Goal is to have a script that manages Steam libraries as a whole, making
sure only one copy of a game is installed across all libraries known to
Steam, checking for game directories that are not listed in any .acf
files, managing special libraries meant for backups or games requiring
updates, etc.

This is a first implementation of copying games to a special library
that is means specifically for games that require updates. It's pretty
quick 'n' dirty at the moment, with the cygwin version of my library
paths hard coded, and no intelligence to cope with a game already in
both places that needs an update to one.