Steam library_sync: Synchronise updates back to main libraries
If the game in the update required library is more recent than the main
libraries, copy it back. Using distutils for the copy operation as it
supports overwriting, and can check the timestamp on the files to
determine if an update is necessary.

Currently refuses to sync back if a game is not in any of the main
libraries or is in multiple libraries, as it is not clear which to
update.