tmux.conf: Use #{pane_current_path} to replace dropped feature in tmux 1.9
Tmux used to have a default-path feature, that would create new windows
with the same working directory as the highlighted pane by default. For
some inexplicable reason this was dropped from tmux 1.9, instead having
to use the -c parameter to explicitly set the working directory.

Someone also forgot to document the #{pane_current_path} feature in the
man page, which gets the current working directory of the highlighted
pane - pass this into -c to emulate the old behaviour.