quickfox: Resend oldest unacknowledged packets first
For the first pass proof of concept I had the script resent a random
outstanding packet when it decided to resend anything while I made sure
everything was working (including correctly dealing with duplicated
resent packets).

This is suboptimal as it might randomly choose a packet that has only
just been sent and hasn't had time the acknowledgement to come in.

Instead, keep the list of outstanding packets in order of oldest to most
recently sent and always resend the oldest packet. On my connection with
0% loss on upload this completely eliminates duplicate packets on the
receiver end so all resent packets were necessary. On links with some
loss in the upload this should also improve things.