i3status: Add available memory display
i3status doesn't support showing free memory because wank wank excuse
excuse blah blah bullshit and the devs have more RAM than they know what
to do with so they don't care.

Add a wrapper script that shows the available memory as reported by the
kernel - this is the number we most care about, because it is the amount
of memory that can be allocated by applications before the kernel will
have to start swapping, or before things start failing if there is no
swap. The display will turn yellow when less than 200MiB are available
and red when less than 100MiB are available.

Enable this display on my Mac Air where I don't use swap and need to
keep this number above zero otherwise things start to go bad - better to
terminate firefox early when it starts becoming a memory hog than be
forced to terminate all of X when the system is otherwise unresponsive.